+++ packages/boilerplate/theme/pages/Checkout/Payment.vue
@@ -310,7 +310,6 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@import "~@storefront-ui/vue/styles";
 .title {
   margin: var(--spacer-xl) 0 var(--spacer-base) 0;
   @include for-desktop {