+++ packages/core/nuxt-theme-module/theme/components/Filters.vue
@@ -140,8 +140,6 @@ export default {
 };
 </script>
 <style lang="scss">
-@import "~@storefront-ui/vue/styles";
-
 .filters {
   &__title {
     --heading-title-font-size: var(--font-size--xl);