+++ packages/commercetools/composables/src/useCheckout/createPlaceOrder.ts
@@ -8,11 +8,13 @@ const createPlaceOrder = ({ cartFields }, customQuery?: CustomQuery) => async ()
   loading.value.order = true;
   const { id, version } = cartFields.cart.value;
 
-  const orderResponse = await createMyOrderFromCart({ id, version }, customQuery);
-  const { order } = orderResponse.data;
-
-  loading.value.order = false;
-  return order;
+  try {
+    const orderResponse = await createMyOrderFromCart({ id, version }, customQuery);
+    const { order } = orderResponse.data;
+    return order;
+  } finally {
+    loading.value.order = false;
+  }
 };
 
 export default createPlaceOrder;