+++ packages/core/cli/__tests__/scripts/createProject/copyAgnosticTheme.spec.ts
@@ -1,26 +1,21 @@
 import copyAgnosticTheme from '../../../src/scripts/createProject/copyAgnosticTheme';
 
-const themePath = '/home/somepath';
+const dependencyPath = '/home/somepath';
 const NOT_EXISTING_PATH = 'not-existing-path-test';
 
 const files = [
-  '/home/somepath/nuxt.config.js',
-  '/home/somepath/package.json',
-  '/home/somepath/components',
-  '/home/somepath/pages',
-  '/home/somepath/assets',
-  '/home/somepath/helpers',
-  '/home/somepath/plugins',
-  '/home/somepath/middleware'
+  '@home/somepackage',
+  '@home/someotherpackage'
 ];
 
 const filesWithNotExistingPath = [
   ...files,
+  '/home/somepath/file.js',
   `/home/somepath/${NOT_EXISTING_PATH}`
 ];
 
 jest.mock('@vue-storefront/cli/src/utils/helpers', () => ({
-  getThemePath: () => themePath,
+  getDependencyPath: () => dependencyPath,
   buildFileTargetPath: (file: string, targetPath: string, chopPhrase: string): string => targetPath + (file.replace(chopPhrase, ''))
 }));
 
@@ -42,7 +37,6 @@ jest.mock('path', () => ({
 
 describe('[vsf-next-cli] copyAgnosticTheme', () => {
   it('compiles & copies template with proper arguments to relative path', async () => {
-
     const integration = 'magento-2';
     const targetPath = '../../my-new-super-project';
 
@@ -52,7 +46,7 @@ describe('[vsf-next-cli] copyAgnosticTheme', () => {
     for (const file of files) {
       expect(compileTemplateMock).toHaveBeenCalledWith(
         file,
-        targetPath + (file.replace(themePath, '')),
+        targetPath + (file.replace(dependencyPath, '')),
         {
           generate: {
             replace: {
@@ -62,12 +56,10 @@ describe('[vsf-next-cli] copyAgnosticTheme', () => {
           }
         }
       );
-
     }
   });
 
   it('compiles & copies template with proper arguments to absolute path', async () => {
-
     const integration = 'magento-2';
     const targetPath = '/home/thecreator/my-new-super-project';
 
@@ -77,7 +69,7 @@ describe('[vsf-next-cli] copyAgnosticTheme', () => {
     for (const file of files) {
       expect(compileTemplateMock).toHaveBeenCalledWith(
         file,
-        targetPath + (file.replace(themePath, '')),
+        targetPath + (file.replace(dependencyPath, '')),
         {
           generate: {
             replace: {
@@ -87,12 +79,10 @@ describe('[vsf-next-cli] copyAgnosticTheme', () => {
           }
         }
       );
-
     }
   });
 
   it('omits not existing paths', async () => {
-
     const integration = 'magento-2';
     const targetPath = '../../my-new-super-project';
     const file = `/home/somepath/${NOT_EXISTING_PATH}`;
@@ -102,7 +92,7 @@ describe('[vsf-next-cli] copyAgnosticTheme', () => {
     await copyAgnosticTheme(integration, targetPath);
     expect(compileTemplateMock).not.toHaveBeenCalledWith(
       file,
-      targetPath + (file.replace(themePath, '')),
+      targetPath + (file.replace(dependencyPath, '')),
       {
         generate: {
           replace: {
@@ -112,7 +102,5 @@ describe('[vsf-next-cli] copyAgnosticTheme', () => {
         }
       }
     );
-
   });
-
 });