+++ packages/core/jest.config.js
@@ -13,8 +13,7 @@ module.exports = {
   moduleFileExtensions: ['ts', 'js', 'json'],
   moduleNameMapper: {
     '^@vue-storefront/cli/(.*?)$': '<rootDir>/cli/$1',
-    '^@vue-storefront/nuxt-theme/(.*?)$': '<rootDir>/nuxt-theme-module/$1',
-    '^@vue-storefront/(.*?)$': '<rootDir>/$1'
+    '^@vue-storefront/nuxt-theme/(.*?)$': '<rootDir>/nuxt-theme-module/$1'
   },
   rootDir: __dirname,
   setupFilesAfterEnv: [