+++ packages/core/nuxt-theme-module/theme/layouts/default.vue
@@ -41,8 +41,6 @@ export default {
 </script>
 
 <style lang="scss">
-@import "~@storefront-ui/vue/styles";
-
 #layout {
   box-sizing: border-box;
   @include for-desktop {