+++ packages/core/nuxt-theme-module/theme/pages/Checkout.vue
@@ -81,8 +81,6 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@import "~@storefront-ui/vue/styles";
-
 #checkout {
   box-sizing: border-box;
   @include for-desktop {