+++ packages/core/docs/commercetools/changelog.md
@@ -3,6 +3,7 @@
 ## 0.2.7 - not released
 
 - [BREAKING] removed `availableFilters` and `availableSortingOptions` from `useProduct` ([#4856](https://github.com/DivanteLtd/vue-storefront/issues/4856))
+- enabled "modern mode" in `yarn build` command ([#5203](https://github.com/DivanteLtd/vue-storefront/issues/5203))
 
 ## 0.2.6